voice_chat/agent/requirements.txt
livekit
livekit-agents
openai
google-api-python-client
google-auth
python-dotenv